call dropdown to displayer
